refactor: Rename `…e/compile/events/api-gateway/lib/hack/updateStage.js`
