+++ lib/plugins/aws/package/compile/events/api-gateway/index.js
@@ -263,7 +263,7 @@ class AwsCompileApigEvents {
         const getServiceState = require('../../../../lib/get-service-state').getServiceState;
 
         const state = getServiceState.call(this);
-        const updateStage = require('./lib/hack/updateStage').updateStage;
+        const updateStage = require('./lib/hack/update-stage').updateStage;
 
         this.state = state;
         return updateStage.call(this);
